+++ src/main/java/zifnab/hdf/DataElement.java
@@ -74,6 +74,18 @@ public final class DataElement
     }
   }
 
+  public void assertTokenCount( final int length )
+  {
+    final int size = _tokens.size();
+    if ( length != size )
+    {
+      final String message =
+        "Data element named '" + getName() + "' expected to contain " + length + " tokens " +
+        "but contains " + size + " tokens";
+      throw new DataAccessException( message, getLocation() );
+    }
+  }
+
   @Nonnull
   public List<DataNode> getChildren()
   {