+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add `DataElement.getChildElements()` helper method that returns a list of the child elements, skipping all the comments.
 * Add the constructor `DataElement(SourceLocation,DataElement,String...)` to simplify manual creation of elements.
 * Add the method `DataElement.assertTokenName(String)` to simplify verification of an element when mapping to configuration objects.
+* Add the method `DataElement.assertTokenCount(int count)` to simplify verification of an element when mapping to configuration objects.
 
 ### [v0.01](https://github.com/realityforge/zifnab/tree/v0.01) (2019-08-06)
 [Full Changelog](https://github.com/realityforge/zifnab/compare/b24bbdea2237c119e17341e5597c42b21b76a9c9...v0.01)